rename: 포인트 컴포넌트 이름 리뷰와 동일하게 수정
